requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.8.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
openai>=1.0.0
